% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgiveness.prep.fund.R
\name{forgiveness.prep.fund}
\alias{forgiveness.prep.fund}
\title{A Forgiveness Tax Bill Function}
\usage{
forgiveness.prep.fund(amount.forgiven, growth.rate, tax.rate, res,
  years = 20, print.saved = 1)
}
\arguments{
\item{amount.forgiven}{Starting value of the loan at time 0.}

\item{growth.rate}{Growth rate of the forgiveness prep fund.}

\item{tax.rate}{Income tax rate.}

\item{res}{The length of your residency.}

\item{years}{The timeframe you want to pay off your debt in. Default is 20.}

\item{print.saved}{Prints saved amount. Default is 1.}
}
\description{
This function allows you to determine how much you need to put away per year
to cover your tax bill on the forgiven amount.
}
\examples{
forgiveness.prep.fund(545732.1, .05, .35, 5, years=20, print.saved=1)
}
\keyword{debt}
\keyword{debt,}
\keyword{forgiveness}
\keyword{tax,}
