% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{grow}
\alias{grow}
\title{A Growth Function}
\usage{
grow(base, interest, n = 1)
}
\arguments{
\item{base}{Starting value of the loan at time 0.}

\item{interest}{Interest rate on the loan.}

\item{n}{The number of times the loan should grow: ex. the number of years
if it is an annual interest rate. Defaults to 1.}
}
\description{
This function allows you to grow a loan.
}
\examples{
grow(100000, .075, n=8)
}
\keyword{growth,}
\keyword{interest}
