% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.driven.R
\name{income.driven}
\alias{income.driven}
\title{A Loan Forgiveness Function}
\usage{
income.driven(base, interest, income, plan = 20)
}
\arguments{
\item{base}{Starting value of the loan at time 0.}

\item{interest}{Interest rate on your loan.}

\item{income}{Income.}

\item{plan}{The timeframe you want to pay off your debt in. Default is 20.}
}
\description{
This function calculates how much of your loan is forgiven.
}
\examples{
income <- c(65000,65000,65000,65000,65000, 353000, 353000,353000, 353000,353000, 353000,353000, 353000,353000, 353000,353000, 353000,353000, 353000,353000)
income.driven(400000,.075,income, plan=20)
}
\keyword{income}
\keyword{interest,}
