% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pay.per.year.R
\name{pay.per.year}
\alias{pay.per.year}
\title{A Pay Per Year Function}
\usage{
pay.per.year(debt.total, interest, res, years = 20)
}
\arguments{
\item{debt.total}{The total amount of debt you have remaining.}

\item{interest}{Interest rate on the loan.}

\item{res}{The length of your residency.}

\item{years}{The timeframe you want to pay off your debt in. Default is 20.}
}
\description{
This function allows you calculate how much you need to pay per year
to pay off your loans after residency.
}
\examples{
pay.per.year()
}
\keyword{debt,}
\keyword{interest}
