% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{payments}
\alias{payments}
\title{A Payments Function}
\usage{
payments(base, interest, payment)
}
\arguments{
\item{base}{Starting value of the loan at time 0.}

\item{interest}{Interest rate on the loan.}

\item{payment}{The annual payment you intend to make.}
}
\description{
This function allows you to keep track of your payments toward your debt.
}
\examples{
payments(100000, .075, 10000)
}
\keyword{debt,}
\keyword{interest}
