% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialty_res.R
\name{specialty_res}
\alias{specialty_res}
\title{A Length of Residency Function}
\usage{
specialty_res(specialty)
}
\arguments{
\item{specialty}{The specialty of interest.}
}
\description{
This function returns the minimum length of residency for a given specialty.
}
\examples{
specialty_res()
}
\keyword{residency,}
\keyword{specialty}
