% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialty_salary.R
\name{specialty_salary}
\alias{specialty_salary}
\title{A Specialty Salary Function}
\usage{
specialty_salary(specialty)
}
\arguments{
\item{specialty}{The specialty of interest.}
}
\description{
This function returns the average attending salary for a given specialty.
}
\examples{
specialty_salary("Internal Medicine")
}
\keyword{residency,}
\keyword{salary}
\keyword{specialty,}
